Move DefaultCapabilities() to caps package
